+++ GrainElevatorAPI.Core/Interfaces/ModelInterfaces/ICompletionReportOperation.cs
@@ -4,7 +4,7 @@ public interface ICompletionReportOperation
 {
     int Id { get; set; }
     int TechnologicalOperationId { get; set; }
-    int Amount { get; set; }
+    double Amount { get; set; }
     decimal? OperationCost  {get; set; }
     int? CompletionReportId { get; set; }
